chore: update commons mod names
