+++ src/cdf/core/adapter/extract_load.py
@@ -17,7 +17,7 @@ from types import ModuleType
 
 import cdf.core.interface as I
 from cdf.commons.file import json, load_module_from_path, yaml
-from cdf.commons.general import inject_sys_path
+from cdf.commons.pyutils import inject_sys_path
 from cdf.core.container import Container
 
 __all__ = ["DltAdapter", "SlingAdapter", "SingerAdapter", "HamiltonAdapter"]
